requirements.txt
fastapi
Pillow
python-multipart
google-generativeai
python-dotenv